+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.0
+
+- Update package dependency.
+- Allow passing `PieSocketOptions` per room.
+- Add ability to listen for events by prefix.
+- Add `closeAll` method to close all existing connection.
+- Explicitly invoke the `connect` method to manage initial error and implement retry logic.
+
 ## 1.0.1
 
 Moves example into private directory, no functional changes.